+++ GUI/include/TTauri/GUI/PipelineImage.hpp
@@ -37,7 +37,7 @@ protected:
     PushConstants pushConstants;
     int numberOfAtlasImagesInDescriptor = 0;
 
-    int numberOfVertices = 0;
+    ssize_t numberOfVertices = 0;
     std::vector<vk::Buffer> vertexBuffers;
     std::vector<VmaAllocation> vertexBuffersAllocation;
     std::vector<gsl::span<Vertex>> vertexBuffersData;
