+++ GUI/include/TTauri/GUI/WindowTrafficLightsWidget.hpp
@@ -45,7 +45,7 @@ public:
 
     int state() const noexcept;
 
-    void pipelineImagePlaceVertices(gsl::span<PipelineImage::Vertex>& vertices, int& offset) noexcept override;
+    void pipelineImagePlaceVertices(gsl::span<PipelineImage::Vertex>& vertices, ssize_t& offset) noexcept override;
 
     void handleMouseEvent(MouseEvent event) noexcept override;
     HitBox hitBoxTest(glm::vec2 position) const noexcept override;
