+++ Widgets/include/TTauri/Widgets/ButtonWidget.hpp
@@ -49,8 +49,8 @@ public:
 
 
     void update(bool modified) noexcept override;
-    void pipelineImagePlaceVertices(gsl::span<GUI::PipelineImage::Vertex>& vertices, int& offset) noexcept override;
-    void pipelineMSDFPlaceVertices(gsl::span<GUI::PipelineMSDF::Vertex>& vertices, int& offset) noexcept override;
+    void pipelineImagePlaceVertices(gsl::span<GUI::PipelineImage::Vertex>& vertices, ssize_t& offset) noexcept override;
+    void pipelineMSDFPlaceVertices(gsl::span<GUI::PipelineMSDF::Vertex>& vertices, ssize_t& offset) noexcept override;
 
     void handleMouseEvent(GUI::MouseEvent event) noexcept override;
